Move BMI code to bmi module
