+++ PyStemmusScope/bmi/docker_process.py
